+++ commands/list.go
@@ -118,7 +118,8 @@ var listSpecificResourceCmd = func(resType string) *cobra.Command {
 			} else {
 				srv, err := cloud.GetServiceForType(resType)
 				exitOn(err)
-				g, err = srv.FetchByType(context.WithValue(context.Background(), "force", true), resType)
+				fetchContext := context.WithValue(context.Background(), "force", true)
+				g, err = srv.FetchByType(context.WithValue(fetchContext, "filters", listingFiltersFlag), resType)
 				exitOn(err)
 			}