+++ config/upgrade_test.go
@@ -38,7 +38,7 @@ func TestUpgradeMessaging(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	exp := fmt.Sprintf("New version 1000.0.0 available. Changelog at https://github.com/wallix/awless/blob/master/CHANGELOG.md\nRun `wget -O awless-1000.0.0.tar.gz https://github.com/wallix/awless/releases/download/1000.0.0/awless-%s-%s.tar.gz`\n", runtime.GOOS, runtime.GOARCH)
+	exp := fmt.Sprintf("New version 1000.0.0 available. Checkout the latest features at https://github.com/wallix/awless/blob/master/CHANGELOG.md\nRun `wget -O awless-1000.0.0.tar.gz https://github.com/wallix/awless/releases/download/1000.0.0/awless-%s-%s.tar.gz`\n", runtime.GOOS, runtime.GOARCH)
 	if got, want := buff.String(), exp; got != want {
 		t.Fatalf("got %s, want %s", got, want)
 	}